track progress of updating apps
refs #8411

+++ framework/src/timeintegrators/LStableDirk4.C
@@ -107,7 +107,9 @@ LStableDirk4::postStep(NumericVector<Number> & residual)
 {
   // Error if _stage got messed up somehow.
   if (_stage > _n_stages)
-    mooseError2("LStableDirk4::postStep(): Member variable _stage can only have values 1-", _n_stages, ".");
+    // the explicit cast prevents strange compiler weirdness with the static
+    // const variable and the variadic mooseError function
+    mooseError2("LStableDirk4::postStep(): Member variable _stage can only have values 1-", (unsigned int)_n_stages, ".");
 
   // In the standard RK notation, the residual of stage 1 of s is given by:
   //
